requirements.txt
Flask>=3.0.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
Pillow>=10.0.0
deepface>=0.0.79
opencv-python>=4.8.0
tf-keras>=2.15.0
gunicorn>=21.0.0
